+++ MRConnectDemoProject/MRConnectDemoProject/Model/Specialities.swift
@@ -9,17 +9,17 @@ import Foundation
 
 struct Specialities {
     
-    static let specialities: [Int16:String] = [
-        0:MyStrings.cardiologist,
-        1:MyStrings.dermatologist,
-        2:MyStrings.gynecologist,
-        3:MyStrings.neurologist,
-        4:MyStrings.oncologist,
-        5:MyStrings.pediatrician,
-        6:MyStrings.physician,
-        7:MyStrings.psychiatrist,
-        8:MyStrings.radiologist,
-        9:MyStrings.surgeon
+    static let specialities: [Int: String] = [
+        0: MyStrings.cardiologist,
+        1: MyStrings.dermatologist,
+        2: MyStrings.gynecologist,
+        3: MyStrings.neurologist,
+        4: MyStrings.oncologist,
+        5: MyStrings.pediatrician,
+        6: MyStrings.physician,
+        7: MyStrings.psychiatrist,
+        8: MyStrings.radiologist,
+        9: MyStrings.surgeon
     ]
     
 }
